requirements.txt
Werkzeug==1.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
click==7.1.2
Flask==1.1.2
kafka-python
python-dotenv
pytest
pytest-mock
pytz